Rewrote the record parsers so they don't allow missing IDs. You can raise an error, report and skip, or simply skip records which have an error.
While I want to allow a bare SMILES string as SMILES input,
it turned out to be difficult in the face of multiple structure
files. That is, my low-level readers were assigning the same
ID to records from different filenames if both were missing
a name, and were at the same order in the file.

I tried a solution where I returned None as the id, and let
the higher-level API assign its default name. While it worked,
I realized that the next problem is that there's no easy way
to map back from that name to the originating record. (You
might think that "Mol_3" goes to record #3 but that's not true
because some of the toolkits simply skip unknown records.)

Therefore, all records must have an identifier.