Added the Gillies method as one of the options. Fixed a bug in my timing probe.
In short, my tests shows that the Gillies method is faster than the
LUT on my desktop, but the probe showed it the other way. I
conjectured that my test fingerprint was too small, and the parts of
the LUT which I was using could fit into L1 cache. I increased the
number of fingerprints and the ranking is now in line with what it
should be.