Fixed a serious bug in how subareana identifiers work. I was using the primary arena identifiers instead of the subarena identifiers. This goes back to a naming decision I made where I implemented the names incorrectly - I thought I had good justification, but the rest of the code as implemented proved me wrong.
Two of the test cases should have caught it (though there should have
been three test cases to catch it). It looks like I believed the
computer more than I believed myself, and commited it with the wrong
test code. I fixed those two cases (for the threshold and knearest
searches) and added one for the count code.

In reviewing the changes to the tests, I realized that there were
related tests I was missing. I needed to check when the subarena
was not also the first subarena of the arena. That is, I never
checked to see if my id offsets were correct.

Now I do.

Thanks to Brian McClain for the testing which identified this problem.