iter_indices_and_scores returned a SearchResult when it should have returned a list of 2-element tuples. This worked because a SearchResult can be iterated to return the 2-element tuples. However, I want it to return a real list.
The tests didn't catch it because I used assertListEquals, which
allows iterators. I ended up making an assertRealListEquals to
require that the chemfp API returns either a list or an array.

Fixing this revealed a bug elsewhere in the tests where I assumed
that iter_indices_and_scores iterates over SearchResult instances.