Removed the commented-out stubs for the partial k-nearest cases.
I couldn't figure out how to do it. Fundamentally the lower-triangle
and upper-triangle are not symmetric, so it's not as simple as the
threshold implementations. There can be some shared information, but
that requires more locking than I wanted to do. I decided that people
will just have to do the full search.